moved resize and frame functions out of draw-map as component graph-frame
axis-2d.js : resizeEffect : alias : insert graphFrame.

draw-map.js :
move some imports with functions / code moved to other files, when no longer required here.
draw() : use setProperties to merge axisApiAdd into axisApi.
moved to graph-frame.js :
  compareViewport()
  setupViewport()
  renderFrame()
  setCssVariable()
  flowName(), flowHidden(), renderForeground()
  recordViewport(), showResize()
  updateSelections()
  updateSyntenyBlocksPosition()
  $.popover setup, readLeftPanelShown()
  drawEffect(), resizeEffect(), stacksWidthChanges(), readLeftPanelToggle(), stacksWidthChanged(), resize()
moved to axisBrush.js : zoomBehaviorSetup()
moved code to graph-frame : renderFrame()
moved code to app/utils/draw/flow-controls.js: updateSelections_flowControls() (new function)
moved to stacks-view.js : splitAxes, stacksRedraw(), axesShowXOffsets()

block-adj.js : updatePathsPositionDebounce{,d}() : stacksWidthChanges has moved from graphFrame to drawMap.

stacks-view.js :
init() : call stacks.init(), use imported stacks reference because this.oa.stacks is not set when first called.
add didInsertElement() : drawMap.draw().
add splitAxesLength.

mapview.js : add to .oa : axisApi, stacks.
draw-map.hbs : use draw/graph-frame. move {{resizeEffect}} {{drawEffect}} to form graph-frame.hbs.
grouping.js : added blocksReferencesBeforeData() using unneeded code from drawEffect().

axisTitle.js : update comment re. splitAxes1 being changed to true by default in f636c80.
chart1.js : blockColour() : reduce trace (this may be occurring initially, before colourSlotsUsed[] is augmented).

stacks.js : stacks.init() does not need to initialise stacks.oa - done in mapview.js : init()
