+++ frontend/app/templates/components/draw-map.hbs
@@ -16,7 +16,12 @@
 {{yield}}
 
 <div class="devel-visible">
-  {{resizeEffect}} {{drawEffect}} 
+  {{draw/graph-frame oa=oa drawMap=this drawMap=this
+    controls=controls
+    componentGeometry=componentGeometry
+    data=this.data
+    panelLayout=panelLayout
+  }}
   {{draw/stacks-view stacksOld=oa.stacks oa=oa drawMap=this axes2d=axes2d
     model=model
     featuresInBlocks=featuresInBlocks
