+++ frontend/app/templates/components/draw/graph-frame.hbs
@@ -1,0 +1,1 @@
+  {{resizeEffect}} {{drawEffect}}
