+++ frontend/app/utils/draw/chart1.js
@@ -1002,7 +1002,9 @@ ChartLine.prototype.blockColour = function ()
   }
 
   if (! blockS || ! blockS.axisTitleColour()) {
-    dLog('blockColour', blockS);
+    if (trace > 1) {
+      dLog('blockColour', blockS);
+    }
   }
   let
   /* For axes without a reference, i.e. GMs, there is a single data block with colour===undefined. */
